// Frontend/src/api/generated/model/createRoleResponse.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { RolePermissionDto } from './rolePermissionDto';

export interface CreateRoleResponse {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  permissions?: RolePermissionDto[] | null;
}