// Frontend/src/api/generated/model/updateUserRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { RolePermissionDto } from './rolePermissionDto';

export interface UpdateUserRequest {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  password?: string | null;
  roleId?: string;
  isActive?: boolean;
  /** @nullable */
  extraPermissions?: RolePermissionDto[] | null;
}