// Frontend/src/api/generated/model/userDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { RoleDto } from './roleDto';
import type { RolePermissionDto } from './rolePermissionDto';

export interface UserDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  phone?: string | null;
  isActive?: boolean;
  role?: RoleDto;
  /** @nullable */
  extraPermissions?: RolePermissionDto[] | null;
}