// Frontend/src/api/generated/model/userMasterDataDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { RoleDto } from './roleDto';
import type { RolePermissionDto } from './rolePermissionDto';

export interface UserMasterDataDto {
  /** @nullable */
  roles?: RoleDto[] | null;
  /** @nullable */
  permissions?: RolePermissionDto[] | null;
}